requirements.txt
numpy
pandas
matplotlib
seaborn
afinn
sqlalchemy
pillow